Added upstream surface
Added the functionality of all upstream surface particles. The methods looks very successful, with vortex shedding occuring, and several realistic phenomena occuring. Another possible test case that comes to mind is a droplet dropping in air, and how it breaks up.
One of the great advantages of this method is it's intelligence to find the particles that are in the flow, meaning that there isnt possibility for unforced particles to slip out.
A lot of work needs to be done however to find a relationship between the force and the particle size, as when the particle size is decreased, experienced force is not constant. This is probably something that needs to be done across all the different options to find a normalised relationship to an onset velocity.
Another step is to look at relating the force to the surface tension value, however this also seems to vary with particle fineness, making that fitting more challenging. The hope was that this could then try and approach an adaptation to surfaces and droplets of different sizes.
Another possible idea to look at is potentially using an aggregate mass from the surrounding neighbours to translate force... Many possiblilities... Which is best?