Added Ghost Particle option
added in the option of ghost particles to transmit the force on the particles. Need to modify to allow them to remain for at least a timestep, so that the velocity can actually translate a force. If the generator can be modified to check for prexisting particles then I can just add them where it needs to go, and delete any that don't fall in a radius of a particle, instead of deleting every timestep. So some work to be done. It also tends to delete simulation particles atm, so much work needs doing (but it isnt seg faulting, so there's that goin for it).