Improvements to stability
I perhaps found an instability waiting to happen... In trying to normalise the aerodynamic force, I found that depending on the size of the force, the program would segfault, and in debugging found it was solved by taking a smaller timestep. There are many other possible ways to alleviate the issue, I suspect, but certainly shortening the timestep is one of the safest. Alternatives are relying on damping from viscosity or artificial viscosity, XSPH correction or surface tension, which may well not be enough. Ultimately it means another dimension of tuning to be done. On a plus, the force is now normalised from the particle radius. I now just need to provide a correction constant (hopefully) to get realistic physics.