Fixed Numerous Bugs
In trying to replicate experimental setups, I discovered a variety of errors in the code.
The first, and non-destructive problem, was the error calculation. I wasn't deleting the state from the previous iteration, meaning that the error was near identical to the original, resulting in no convergence. Easy fix on line 878.
The second problem was that of the surface tension. This manifestation was first identified in the timestep, which went to 0, because of the variable timestep calculation. The surface tension force was absolutely huge, tending the max force to infinity. Since dt is proportional to this value inversely, it meant that dt -> 0. It transpired that this was because the contribution was divided by particle mass twice, instead of the correct singular. So now the force isn't going to infinity, when output to the terminal. However when output is suppressed, the value goes to infinity (I added a catch to exit if this is so)... The 'hack' (since there is certainly no understanding present) was to turn on the -ffast-math flag, and for some reason it works!
Still working on the ghost-particle side of things, but trying to find a solid way to replicate experimental data at present.