Updated input format
Several further updates have been made, primarily to the input of the code. Further issues with instability have been found, however they're related to the aerodynamic forces that are being researched. It's possible that something to handle cases of particularly high force are necessary, however for now I just need to be careful.