Updates to timestep + another Aero force translation
A new aero force translation has been implemented, looking at using a correction factor based on the angle formed from the surface to the aerodynamic vector. Seems promising. Has some great advantages such as capability for shear on parallel airflow, as well as blowback on the backside of the jet. Also made the timestep adaptive to ensure a stable timestep each time.