Added particle data reserve
Added a function to calculate how many particles there are going to be in the simulation, and then reserve that space for the vectors.
I may modify the whole data structure in future to be able to use even more efficient memory initialisation, but for now this is definitely the best memory solution available, without fancy initialisation in the Particle class.