SPHERIC update and improvement
So this is a pretty significant update from the previous code that has been on the VCS, and several things have been streamlined. The 3D needs to have some fat cut from it, but this has already been done with the 2D code. This is also adding the 3D side of things to the repository, now that development is shifting to 3D. 2D is still being refined as it's a much quicker way to test things. 
Several errors have been found and corrected, such as surface tension, and how aerodynamic force is applied. Many things have also been dropped, not that they were particularly supported, most notable of which is ghost particles. These will not be pursued further and have been removed from the code, and shall be gone soon from the 3D code. All surface particles has also been dropped since it was terrible, and pointless. The remaining aerodynamic translations are now: 
Left half
Upwind particles
all surface proportional to aero-surface angle.
point vortex.
surface -aero angle, with additional correction based on number of neighbours (Latest testing case).

The input files need to be changed to match these new cases as the numbers have been reordered, which shall follow shortly.