Removing XSPH Correction + New ST model
Removing the XSPH correction as it appears to introduce instability in the 3D code, and its also not necessary for the use application. Also of significant note is the option of another surface tension model, based off of He et al (2014). It claims to be better for thin features, and this may be so but it appears to be a bit more noisy in comparison, though I need to do more comparison. Little information is also available on the factor kappa in the model, and its value for fluids.