Large change to VLM case
Added a large amount to the VLM case allowing sweep and flaps to be specified, as well as an input file to control it.
The jet case now can be specified arbitrarily in space and rotation, but still is circular.
Output has also been modified to create a folder in the Outputs folder, and all files are placed inside here to give a cleaner appearance. I need to modify so that the optional third argument is the name of the folder, so that multiple runs can be done without overwriting.