Added tapered nozzle option
A few changes, but relatively minor, have gone on. I've added an option in svar.Bcase == 2 to have a tapering nozzle. It was hoped that this would have an effect in changing the spray out of nozzles, but it doesn't appear to have a great change.
Iterators have also been changed to be ii, jj, kk etc as opposed to i,j,k throughout.