Working cell containment
The cell containment appears to be reliable in it's operation, at last. The amount of neighbours to seach through appears to be a sensitive variable, and this may be pushed up again to ensure a proper robustness. More debug outputs are slowly being added to help with the process and break each section down.