Cell data restructure
A restructure of the cell data has occured, to take it into a fully face based system. Hopefully this should save some memory requirement, since now there is simply a list of faces, and an array of indexes to those faces.
This is somewhat in reponse to how the check for internal boundaries has worked out. The check is done by seeing if a ray that is cast from the particle to the cell centre crosses a boundary face. If this is true, then it lies on the other side of that face to the cell, i.e. outside of the mesh.