Adding check for boundary
Working on adding checks to see if a particle has crossed a boundary, which appear to be unsuccessful, but it's closer to hopefully working. Work has also been done on an edge based data structure for the 2D cases to get the same information about internal boundaries.
A boundary layer profile has also been added to the fuel exiting the pipe. Some investigation should probably be done to see if this changes the breakup characteristics at all, but hopefully should make things somewhat more realistic particularly because of the destruction of an ordered grid for the particles.