Working boundary handling
The handling of encountering boundaries is hopefully sorted. Some remnants of particle types proved to be stopping regular cell checks. Internal boundaries still need to be verified, and a proper treatment of  the splash also potentially implemented. At present, it's just reflecting the velocity component about the plane of the boundary.
The variable type 'ldouble' has also been changed to 'real' to allow use of either float or double, or possibly even long double (needs testing).