Input Rearrange
A large rearrangement of the input and output has happened, which aligns it closer to a common solver. The output folder is now a subdirectory in the case folder. It is intended to modify the code to take a single input file that has variable name first, and can allow multiple different runs in the same folder, saving duplicates of meshes and solutions.