Fixed restart function
I found a subtle bug in the restart function where I forgot to initialise the partID in each particle structure. This results in no forces being computed as it thinks all particles are the same one. Also modified it so that time output for the boundary is also needed, to save possible instability.