Updated Momentum Exchange
Unfortunately haven't been maintaining updates as frequently as I would like. Many alterations have been made to the code, and the most outstanding of which is the refined momentum exchange from the cell to the fuel. Now, the momentum imparted on the fuel is taken out of the cell. Care needs to be taken however to avoid cells of similar volume to the particles. In this case, the calculation is highly sensitive (More work to be done there). The method works reasonably with a coarse mesh, and some dissipation is present.
Other code changes are a movement from west const, to east const to make a better code style.