Implemented particle shifting
Since the kernel method creates a ton of tensile instability, I implemented particle shifting and tensile instability control, and converted the momentum equation to be completely the same as delta+ SPH.
I had trouble with particles then sticking to the boundary, but this was fixed by providing a proper boundary layer, 4 particles deep. Now flow is considerably smoother in the pipe regions.