Fixed sign issue
Been having some issues with signs when trying to convert to to the delta+ version of SPH. Instead of trying to put it into my own sign convention, I tried to shift mine to the papers. It's now fully in my own sign work, and works as expected. It does what I hoped and has successfully controlled tensile instability, that originally I thought was an error I introduced, but just how normal SPH solves. Momentum is still the monaghan formulation, as I think it is a better pressure gradient, but more testing should be done to confirm, as the d+ version is a bit quicker. I'm about to experiment with trying to take the neighbour search and L matrix computation out of the inner NB loop, by doing the first iteration to find the preliminary n+1 state, and using that to do the search before freezing.
As such, the integration has been moved into a new function, as well as error calculation.