Working Jet case
After much trying, I realised that the L matrix was the negative of what it should be, and so the eigenvalues were all negative. This means that the TIC and PST never kicked in, and now that I've rectified the surface identification, neither are remotely stable. As such, they have been commented out, and hopefully the surface noise is not too bad. We shall wait and see after running some simulations.
The calculation of the kernel sums and the neighbour search however have successfully been removed from the integration loop, saving a lot of computation.