Working surface
The last commit ended up with the same results as old code, with unstable surfaces, possibly because of artificial viscosity, or a poor kernel gradient. The L matrix has been applied now to the gradient, and this seems to have helped to fix the surface. It also seems to make the TIC stable inside the pipe, so I've kept it on, however the particle shifting is just unstable however which way I try at the surface. In the fluid, motion is stable, however the surface is totally unsuccessful, no matter how I attempt to modify the equations.