Fixed Error calculation
Fixed a small error in the newmark scheme where the error I was using to check the while loop wasn't being updated, causing it to always do the maximum iterations. Now it's being updated correctly and exiting. However, the error is a bit debatable as to the correct value, since most are shown as inf to finish.