Improved Pressure integral formulation
The pressure integral formulation was struggling to be stable due to a variety of reasons, with force exploding, thus making the timestep uselessly small. Since the force is highly discontinuous across the pipe exit, a smoothing kernel is added to acclimate the fluid to the aerodynamic force. This is based from the centre of the pipe, with a support radius of the pipe diameter. The means there is still discontinuity at the pipe edges, however the flow appears to be stable so far. If it isn't, then it will need adjusting to the nearest point on the plane.
The delta-SPH component to continuity also had a potential to divide by zero, and eps*h^2 has been added to mitigate, since some points were going to nan position.