Working particle shifting
Finally got a stable version of the particle shifting implemented, as well as an enhanced Cp function, that considers how Cp is for a flat plate. The function interpolates based off of curvature between each of the distributions.
The fluid file has also had a modification, no longer requiring the Newmark Beta parameters or the smoothing kernel factor.