Major March Update
Various changes that can't be summed up with a single title.
Realised the kernel correction factor was incorrect, amongst a variety of incorrect maths.
Particle shifting and ALE-SPH is still very much in testing. I haven't managed yet to get a convincingly stable simulation. The surface becomes highly disordered and may well make the equations tricky.
I also adjusted the input files to change units. The starting coordinates of the jet are now in the grid units, but everything else is strictly in meters.