Improved restart filing
Restart functions now rely on the TECIO library, and this is used so that I can be assured of retention of the double values for things such as the spacing, as currently restarting longer simulations appears to be unstable. The cause is unknown, however the force is enough to cause particles to become distant from each other in the first iteration, suggesting that something is not appropriately initialised. This now means that TECIO must be present and it may be necessary to provide an install of the library in the makefile. That is not going to be done soon however, if at all, versus trying to explore a way that does not require it.