Implemented particle tracking
Functionality has been added to transition to particle tracking, and save computation. Presently this is purely based on reaching a certain distance in the x-direction, however it would be more sensible in future to base this off of some measure of the neighbourhood density, so that only droplets are converted, and not particles that represent a continuum. 
A few other bug fixes have been done alongside.