Improved stability controls
A number of additions have been implemented, the first being a limiter option for the density variation. This helps to prevent simulations going out of control, as some were observed to have small areas 'pop' and explode presumably due to divergence of density. Control is still advised with the speed of sound, but this provides a numerical bounding for stability, and should be imposed only on unphysical densities. 
A new library has also been included to perform a drag sweep for a spherical droplet, and benchmark the coupling forces. 
A bug has also been fixed where a check for restart files is done before the boundary setting is defined, breaking restarts in simulations that include the boundary. It should be intended at some point to write simulation settings to the auxiliary data of the output files, and read key settings from there (such as frame timings and particle spacing), should the para files differ.