Updated Surface tension + IO
Updated the surface tension model to replicate Joe De Courcy's formulation of the Pairwise model, which is better in a number of ways. It's fully localised and also accounts for the density variations that occur in a weakly compressible formulation. 
IO has also been changed to check for parameter equality, rather than use a string.find() function. This was causing issues where I happened to name density variation parameters to have the same sentence subset. Now it checks that the parameter is exactly the parameter definition, rather than simply containing it, however the density variation parameter definition has also been changed.