Improved face converter
Improved the face conversion software to take minutes instead of hours. Not sure why it was so slow, but the new algorithm, taken from the TAU to zCFD python script has drastically improved the efficiency. Some code cleanup has also been done, and a tecplot binary file option added for debugging. Can be commented out if access to TECIO is not available.