Upload dam break validation cases
Two options of dam break have been uploaded. The first being the n2 = 1, being the unit square dam break case in the original Monaghan 1992 paper that uses data from Martin and Moyce. 
Second is used in the delta-SPH by Marrone et al. (2011), and is a 2x1m block of fluid, with the box x-dimension being 5.367m. A high wall is needed to fully contain the fluid, without particles sticking to the boundary and circling back.