Fixed cylinder and inlet inputs
Major update is a fix to the input configurations and definition of the cylinder and inlet geometry types, particularly in 3D. 2D needs testing again, but it expected to work nicely, since it's a lot simpler. Deletion zones also need testing as to whether they are successful.
Terminal output has also been adjusted to show more useful information, such as timestep residual and iteration count, alongside the other information present.
Slowly trying to eliminate the use of iostream, since it is known to be abhorrently slow, but now it's only used in non time critical areas for the most part.
VLM has now also been transitioned to a .cpp and .h format.
Adjusted the makefile to place object files in the build directory rather than the source directory to save uploading to onedrive every time. Need to add option for cleaning object files only though.