Fixed issues with 2D cylinder and inlet geometries
Some issues have been remedied with 2D cylinder and inlet geometries, as well as changing the default normal to the positive x-axis, which will then be aligned with the base freestream velocity, making more sense.
Some adjustments to the VLM code as well to not break when compiling 2D.
Also created a doc folder to hopefully contain the documentation for the code.