Fixed circle/sphere geometry, normal directions and curvature
Fixes to address the circle and sphere geometries to be an actual circle, as opposed to the cube as before.
Also noticed issues with normal directions and occlusion factors altering the drag force, that have been fixed. Similarly curvature is not well calculated for particles with lambda < 0.7, due to a thin feature check, that has been eliminated to have the same calculation for all particles. This should improve the induced pressure representation.
A simulation has also demonstrated the need for adaptation of the CFL number, or more so the importance of converging the timestep to have accurate physics, and thus the CFL is adjusted to attempt to keep convergence. This has been clamped to prevent runaway conditions if a simulation is excessively (read falsely) stable, or unstable. As such more user inputs are available to adjust the criteria around this adaptation.