Added better output support
Output has been improved, and multi-file output is now possible and been tested. The h5part output for multiple files doesn't seem to work quite right (paraview doesn't seem to read it correctly), but I'm unsure as to how to fix it, so its going to sit here for a little bit. It may need a rejig of file naming, which I hope it doesn't require.
Also removed restart functions that are obsolete to clean code up.