Update for new nanoFlann API
A new API for the nanoflann library has been made that has invalidated some of the code here. Some refactoring has been made after realising the amount of code repetition that was affected by such a change, and made some further changes to make a later change less impactful.
Add a few other changes such as being able to set the number of threads that the program will use.