Try improve object orientation
In an effort to try and leverage more of the C++ benefits of objects, a motion to bring things towards objects is being made. It's been semi successful in the direction I tried, with some homologation of the shape functions into namespaces but with consistent function names.
Also been running a formatter on the code to try and introduce some consistency, and hopefully later I will homologate variable and function naming to use a consistent standard.