Add JSON file support
A new potential input method for fluid and boundary blocks has been added, being JSON format. This is a pleasant and scalable file format that allows great flexibility in definition, and is a widespread format.
It also is signalling a bit of the transition into object oriented code. A few other bits of formatting have been sorted, grouping data by type among others.