Change variable ingestion to use a map
Realised that there are new and fancy constructs in C++ that mimic some powerful python structures like a dictionary and these can be effectively used to shorten the code for checking which variables to output. While it has been tested to build, it has not been run and definitely has the possibility to change behaviour.
Also figured out the source of why Var.h was failing in the formatter and it is now correctly formatted.