Deprecate droplet and speed test
Lots of changes but the most major is the removal of droplet and speedtest files as these are not longer correct and are irrelevant code to maintain. Lots of other refactoring going on as well as testing of the new JSON input that is getting improved but still not working correctly for vectors.
Some rearrangement of Eigen libraries to try get linting to work better.