Fix Newmark-Beta integration, refactor Runge-Kutta
Fix an issue with the beta parameter for newmark beta that mean position doesn't update correctly, and error presents as infinite.
Also refactored the Runge Kutta code to avoid repetition and discovered a number of inconsistencies with literature and out of sync forces/positions potential. Not tested yet but for a future commit.