Refactor variable names in svar. Remove ghost code.
Finally refactor to have a consistent pep-8 like standard for variable names in svar. Other parameters have been affected, but fvar and avar need to be adjusted in another commit. Also fixed Runge-Kutta integration that would fail due to Af being empty. RMS error does not appear to be correct but need to do some more work to identify why and how. I wonder if it's to do with half steps for the logbase.
Also fix the conversion program codes.