Refactor to use overriding functions
Refactor the geometry creation to use override of class functions, that inherit the base check_input function. This may slightly change the behaviour because of modifications of the particle spacing and it's knock on to other factors like density and pressure, but this is difficult to manage.
This has a number of virtues though in each class inheriting the variables of the base, so there isn't a need to pass it in.