Refactor to have a common settings class
Refactor the code so that all simulation settings are stored in a single data class. This contains some subclasses to further categorise things, but it means less arguments passing into various functions. There is a debate to make it a global data structure, since it's size is fixed and it should be accessible everywhere. Threading needs to be considered for this however.